requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.1
h11==0.14.0
idna==3.4
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
twilio==7.15.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
